Consider primary key expiration in keyInfo
