+++ lib/key/info.js
@@ -8,7 +8,7 @@ import { createMessage } from '../message/utils';
 async function createPacketInfo(packet, subKey) {
     return {
         algorithm: openpgp.enums.publicKey[packet.algorithm],
-        expires: await subKey.getExpirationTime('encrypt_sign')
+        expires: await subKey.getExpirationTime()
     };
 }
 
@@ -39,23 +39,6 @@ export const getSHA256Fingerprints = (key) => {
             return arrayToHexString(await SHA256(keyPacket.writeForHash(keyPacket.version)));
         })
     );
-};
-
-/**
- * Check whether a key can successfully encrypt a message.
- * This confirms that the key has encryption capabilities, it is neither expired nor revoked, and that its key material is valid.
- * @param {OpenPGPKey} publicKey - key to check
- * @param {Date} date - use the given date instead of the server time
- * @returns {Boolean}
- */
-export const canKeyEncrypt = async (publicKey, date = serverTime()) => {
-    try {
-        await openpgp.encrypt({ message: createMessage('test message'), publicKeys: publicKey, date });
-        return true;
-    } catch (e) {
-        console.log(e);
-        return false;
-    }
 };
 
 const primaryUser = async (key, date) => {
@@ -101,7 +84,7 @@ export default async function keyInfo(rawKey, email, expectEncrypted = true, dat
         created: keys[0].getCreationTime(),
         algorithm: openpgp.enums.publicKey[algoInfo.algorithm],
         algorithmName: algoInfo.algorithm,
-        expires: await keys[0].getExpirationTime('encrypt_sign').catch(() => null),
+        expires: await keys[0].getExpirationTime(),
         encrypt: await packetInfo(await keys[0].getEncryptionKey(undefined, date).catch(() => {}), keys[0]),
         sign: await packetInfo(await keys[0].getSigningKey(undefined, date).catch(() => {}), keys[0]),
         decrypted: keys[0].isDecrypted(), // null if public key
