+++ lib/key/utils.js
@@ -1,6 +1,7 @@
 import { openpgp } from '../openpgp';
 import { serverTime } from '../serverTime';
 import { DEFAULT_OFFSET } from '../constants';
+import { createMessage } from '../message/utils';
 
 // returns promise for generated RSA public and encrypted private keys
 export function generateKey({ passphrase, date = serverTime(), offset = DEFAULT_OFFSET, ...rest }) {
@@ -68,6 +69,22 @@ export async function isExpiredKey(key, date = serverTime()) {
 export async function isRevokedKey(key, date = serverTime()) {
     return key.isRevoked(null, null, date);
 }
+
+/**
+ * Check whether a key can successfully encrypt a message.
+ * This confirms that the key has encryption capabilities, it is neither expired nor revoked, and that its key material is valid.
+ * @param {OpenPGPKey} publicKey - key to check
+ * @param {Date} date - use the given date instead of the server time
+ * @returns {Boolean}
+ */
+export const canKeyEncrypt = async (publicKey, date = serverTime()) => {
+    try {
+        await openpgp.encrypt({ message: createMessage('test message'), publicKeys: publicKey, date });
+        return true;
+    } catch (e) {
+        return false;
+    }
+};
 
 export async function compressKey(armoredKey) {
     const [k] = await getKeys(armoredKey);
