+++ lib/pmcrypto.js
@@ -41,6 +41,7 @@ export {
     getPreferredAlgorithm,
     isExpiredKey,
     isRevokedKey,
+    canKeyEncrypt,
     compressKey,
     getFingerprint,
     getMatchingKey,
@@ -68,7 +69,7 @@ export {
 
 export { parseMail } from './message/processMIME';
 
-export { default as keyInfo, canKeyEncrypt, getSHA256Fingerprints } from './key/info';
+export { default as keyInfo, getSHA256Fingerprints } from './key/info';
 
 export { keyCheck } from './key/check';
